+++ config/dev.exs
@@ -4,6 +4,3 @@ config :uart,
 
 config :farmbot,
   state_path: "/tmp/state"
-
-config :json_rpc,
-  transport: Mqtt.Handler
