+++ lib/farmbot/bot_state/state_debugger.ex
@@ -1,0 +1,24 @@
+defmodule StateDebugger do
+  @moduledoc """
+    Delete me
+  """
+
+  use GenEvent
+
+  def handle_event({:dispatch, state},_) do
+    {:ok, state}
+  end
+
+  def handle_call(:state, state) do
+    {:ok, state, state}
+  end
+
+  def state do
+    GenEvent.call(BotStateEventManager, __MODULE__, :state)
+  end
+
+  def start do
+    Farmbot.BotState.Monitor.add_handler(__MODULE__)
+    {:ok, self()}
+  end
+end
