+++ lib/farmbot/rpc/transport/mqtt.ex
@@ -1,4 +1,4 @@
-defmodule Mqtt.Handler do
+defmodule Farmbot.RPC.Transport.Mqtt do
   require GenServer
   require Logger
 
@@ -18,93 +18,6 @@ defmodule Mqtt.Handler do
     GenServer.start_link(__MODULE__, args, name: __MODULE__)
   end
 
-  def handle_call(_, _, {client, nil}) do
-    Logger.debug("MQTT Doesn't have a token yet.")
-    {:reply, :no_token, {client, nil}}
-  end
-
-  def handle_call({:connect, _message}, _from, {client, token}) do
-    {:reply, :ok, {client, token}}
-  end
-
-  def handle_call({:connect_ack, _message}, _from, {client, token}) do
-    options = [id: 24756, topics: ["bot/#{bot(token)}/from_clients"], qoses: [0]]
-    spawn fn ->
-      Logger.debug("Subscribing")
-      Mqtt.Client.subscribe(client, options)
-      Farmbot.Sync.sync
-    end
-    keep_connection_alive
-    {:reply, :ok, {client, token}}
-  end
-
-  def handle_call({:publish, _message}, _from, {client, token}) do
-    {:reply, :ok, {client, token}}
-  end
-
-  def handle_call({:subscribed_publish, message}, _from, {client, token}) do
-    Map.get(message, :message) |> Poison.decode! |>
-    RPC.MessageManager.sync_notify
-    {:reply, :ok, {client, token}}
-  end
-
-  def handle_call({:subscribed_publish_ack, _message}, _from, {client, token}) do
-    {:reply, :ok, {client, token}}
-  end
-
-  def handle_call({:publish_receive, _message}, _from, {client, token}) do
-    {:reply, :ok, {client, token}}
-  end
-
-  def handle_call({:publish_release, _message}, _from, {client, token}) do
-    {:reply, :ok, {client, token}}
-  end
-
-  def handle_call({:publish_complete, _message}, _from, {client, token}) do
-    {:reply, :ok, {client, token}}
-  end
-
-  def handle_call({:publish_ack, _message}, _from, {client, token}) do
-    {:reply, :ok, {client, token}}
-  end
-
-  def handle_call({:subscribe, _message}, _from, {client, token}) do
-    {:reply, :ok, {client, token}}
-  end
-
-  def handle_call({:subscribe_ack, _message}, _from, {client, token}) do
-    Logger.debug("Subscribed.")
-    spawn fn ->
-      Farmbot.Logger.log("Bot is online and ready to roll", [:ticker], ["BOT STATUS"])
-    end
-    {:reply, :ok, {client, token}}
-  end
-
-  def handle_call({:unsubscribe, _message}, _from, {client, token}) do
-    {:reply, :ok, {client, token}}
-  end
-
-  def handle_call({:unsubscribe_ack, _message}, _from, {client, token}) do
-    {:reply, :ok, {client, token}}
-  end
-
-  def handle_call({:ping, _message}, _from, {client, token}) do
-    {:reply, :ok, {client, token}}
-  end
-
-  def handle_call({:disconnect, _message}, _from, {client, token}) do
-    {:reply, :ok, {client, token}}
-  end
-
-  def handle_call({:pong, _message}, _from, {client, token}) do
-    {:reply, :ok, {client, token}}
-  end
-
-  def handle_call(thing, _from, {client, token}) do
-    Logger.debug("Unhandled Thing #{inspect thing}")
-    {:reply, :ok, {client, token}}
-  end
-
   # dont allow emits when we dont have a token
   def handle_cast(_, {client, nil}) do
     {:noreply, {client, nil}}
@@ -114,7 +27,7 @@ defmodule Mqtt.Handler do
     options = [ id: 1234,
                 topic: "bot/#{bot(token)}/from_device",
                 message: message,
-                dup: 1, qos: 1, retain: 0]
+                dup: 1, qos: 1, retain: 0 ]
     spawn fn -> Mqtt.Client.publish(client, options) end
     {:noreply, {client, token}}
   end
@@ -175,7 +88,7 @@ defmodule Mqtt.Handler do
   end
 
   defp build_last_will_message do
-    JsonRpc.Parser.log_msg("Bot going offline", [:error_ticker], ["ERROR"])
+    Farmbot.RPC.Handler.log_msg("Bot going offline", [:error_ticker], ["ERROR"])
   end
 
   defp login(token) when is_map(token) do
@@ -198,5 +111,97 @@ defmodule Mqtt.Handler do
                will_message: build_last_will_message,
                will_qos: 0,
                will_retain: 0]
+  end
+
+  ##                                ##
+  #  Callbacks from the Mqtt_Client  #
+  ##                                ##
+
+  def handle_call(_, _, {client, nil}) do
+    Logger.debug("MQTT Doesn't have a token yet.")
+    {:reply, :no_token, {client, nil}}
+  end
+
+  def handle_call({:connect_ack, _message}, _from, {client, token}) do
+    options = [id: 24756, topics: ["bot/#{bot(token)}/from_clients"], qoses: [0]]
+    spawn fn ->
+      Logger.debug("Subscribing")
+      Mqtt.Client.subscribe(client, options)
+      Farmbot.Sync.sync
+    end
+    keep_connection_alive
+    {:reply, :ok, {client, token}}
+  end
+
+  def handle_call({:subscribed_publish, message}, _from, {client, token}) do
+    Map.get(message, :message) |> Poison.decode! |>
+    RPC.MessageManager.sync_notify
+    {:reply, :ok, {client, token}}
+  end
+
+  def handle_call({:subscribe_ack, _message}, _from, {client, token}) do
+    Logger.debug("Subscribed.")
+    spawn fn ->
+      Farmbot.Logger.log("Bot is online and ready to roll", [:ticker], ["BOT STATUS"])
+    end
+    {:reply, :ok, {client, token}}
+  end
+
+  def handle_call(thing, _from, {client, token}) do
+    Logger.debug("Unhandled Thing #{inspect thing}")
+    {:reply, :ok, {client, token}}
+  end
+
+  ###                                                   ###
+  #  THESE ARE ALL UNUSED CALLBACKS FROM THE MQTT_CLIENT  #
+  ###                                                   ###
+  def handle_call({:unsubscribe, _message}, _from, {client, token}) do
+    {:reply, :ok, {client, token}}
+  end
+
+  def handle_call({:unsubscribe_ack, _message}, _from, {client, token}) do
+    {:reply, :ok, {client, token}}
+  end
+
+  def handle_call({:ping, _message}, _from, {client, token}) do
+    {:reply, :ok, {client, token}}
+  end
+
+  def handle_call({:disconnect, _message}, _from, {client, token}) do
+    {:reply, :ok, {client, token}}
+  end
+
+  def handle_call({:pong, _message}, _from, {client, token}) do
+    {:reply, :ok, {client, token}}
+  end
+
+  def handle_call({:connect, _message}, _from, {client, token}) do
+    {:reply, :ok, {client, token}}
+  end
+  def handle_call({:publish, _message}, _from, {client, token}) do
+    {:reply, :ok, {client, token}}
+  end
+  def handle_call({:subscribed_publish_ack, _message}, _from, {client, token}) do
+    {:reply, :ok, {client, token}}
+  end
+
+  def handle_call({:publish_receive, _message}, _from, {client, token}) do
+    {:reply, :ok, {client, token}}
+  end
+
+  def handle_call({:publish_release, _message}, _from, {client, token}) do
+    {:reply, :ok, {client, token}}
+  end
+
+  def handle_call({:publish_complete, _message}, _from, {client, token}) do
+    {:reply, :ok, {client, token}}
+  end
+
+  def handle_call({:publish_ack, _message}, _from, {client, token}) do
+    {:reply, :ok, {client, token}}
+  end
+
+  def handle_call({:subscribe, _message}, _from, {client, token}) do
+    {:reply, :ok, {client, token}}
   end
 end
