+++ lib/json_rpc/parser.ex
@@ -1,6 +1,5 @@
-defmodule JsonRpc.Parser do
-  require Logger
-  # notification
+defmodule RPC.Parser do
+  # Notification
   def parse(%{"id" => nil, "method" => method, "params" => params}) do
     %Notification{id: nil, method: method, params: params}
   end
@@ -10,50 +9,12 @@ defmodule JsonRpc.Parser do
     %Request{id: id, method: method, params: params}
   end
 
+  # Response
   def parse(%{"result" => result, "error" => error, "id" => id}) do
     %Response{result: result, error: error, id: id}
   end
 
   def parse(_blah) do
     :not_valid
-  end
-
-  # JSON RPC RESPONSE
-  def ack_msg(id) when is_bitstring(id) do
-    Poison.encode!(
-    %{id: id,
-      error: nil,
-      result: %{"OK" => "OK"} })
-  end
-
-  # JSON RPC RESPONSE ERROR
-  def ack_msg(id, {name, message})
-    when is_bitstring(id) and is_bitstring(name)
-     and is_bitstring(message)
- do
-    Logger.error("RPC ERROR")
-    Logger.debug("#{inspect {name, message}}")
-    Poison.encode!(
-    %{id: id,
-      error: %{name: name,
-               message: message },
-      result: nil})
-  end
-
-  @doc """
-    Logs a message to the frontend.
-  """
-  def log_msg(message, channels, tags)
-  when is_list(channels)
-       and is_list(tags)
-       and is_bitstring(message) do
-    Poison.encode!(
-      %{ id: nil,
-         method: "log_message",
-         params: [%{ status: %{location: Farmbot.BotState.get_current_pos}, # Shhhh
-                     time: :os.system_time(:seconds),
-                     message: message,
-                     channels: channels,
-                     tags: tags }] })
   end
 end
